Add integration hub interactions
The source of data was a mapping created from the Hub's database.
Details: https://dsdmoj.atlassian.net/browse/TA-265 (closed link)

The interactions documented here are massively simplified by type.

There are multiple jobs which extract user roles; the diagram only shows
a single arrow. Each Hub job does something specific so I did not feel
it important to distingish between them at this level.

It may be more important to stop displaying the "Hub" as a system and
treat it as a messaging bus and show the target system instead.