Add container diagram for CCMS provider user interface
I'm deliberately not showing any more of EBS than I need to. This is
because I have very little understanding of EBS at the moment, and don't
fully understand how it is accessed via AccessGate and OAM. For now I'm
only going to show the architecture of provider user interface
and what happens when a provider uses it.

Oracle access manager is part of portal, so just show it as portal on
this diagram and mark it as an external system.

Show the relationship between Oracle HTTP Server, portal and provider
UI. This uses mod_osso and the official docs describe this interaction
in more detail:
https://docs.oracle.com/cd/E52734_01/oam/AIAAG/GUID-42CD58AB-356E-4D3B-B403-7993CC4E6A77.htm#AIAAG1810

The OPA section may be oversimplified. 'OPA' refers to a handful of
Oracle components including OWD, and I don't have a detailed
understanding of how they interact at the moment, but this is my current
mental model of it.

I also don't know the details of the shared schema between EBS/Provider
UI; I've just been told it's a thing, and Capgemini manage this part of
the DB, so we don't have the code for it. See discussion in
https://github.com/ministryofjustice/laa-ccms-pui/pull/18.