Add Portal Context Diagram
We do not currently have a Portal context diagram showing all of the
applications and users that interact with the Portal. This diagram will
help us and others to understand the Portal.

Users have been separated into provider users, internal users and users
who do use the Portal admin screens to reset other users passwords.

Application relationships show how they are integrated with the Portal
and what technology is used to integrate them.