Add Civil Legal Aid reporting landscape
I wanted to capture the overview of Civil Legal Aid reporting:

One of our CLA systems produces a complete dump of data in a zip file.

This file is then downloaded, then subsequently uploaded to another
system by a person.

The target system transforms the data into "subject areas" for
analytical and reporting use and stores it in a database.

The database is used as a data source in our business intelligence
system.