Clarify Authentication/Authorisation relationship when using SAML
On the basis that for CCR, CCLF and Apply we use SAML for both
Authentication and Authorisation purposes, it makes sense to have
just one relationship.

In details, quoting Dave:

CCR and CCLF are passed the role information in the SAML response
(the same as Apply), the application then uses this for authorisation.

Eforms only does Authentication via SAML and does its authorisation by
having a list of eforms users and their roles synced from CWA via the hub
to eforms, eforms then checks the username provided by the portal against
this to provide the relevant permissions. (CCMS and EMI also have user
and role details copied to them via the HUB)